pyproject.toml
[project]
name = "tecoquest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
    "typer>=0.15.2",
]
